Update flèches sur le canvas

+++ Plugin ObjectAid/src/main/java/org/example/sae301objectaid/Model.java
@@ -67,7 +67,7 @@ public class Model implements Sujet {
         }
     }
 
-    public void ajoutFleche(int x1, int y1, int x2, int y2){
+    public void ajoutFleche(double x1, double y1, double x2, double y2){
         this.dessinFleches.add(new DessinFleche(x1, y1, x2, y2)) ;
     }
